adding handle key down mehtod to prevent non numeric values
