+++ apps/modernization-ui/src/design-system/input/numeric/NumericInput.tsx
@@ -2,22 +2,10 @@ import { ChangeEvent as ReactChangeEvent, useEffect, useMemo, useState } from 'r
 import classNames from 'classnames';
 import { EntryWrapper, Orientation, Sizing } from 'components/Entry';
 import { Numeric } from './Numeric';
-import { KeyboardEvent as ReactKeyboardEvent } from 'react';
 
 type NumericOnChange = (value?: number) => void;
 
 const asDisplay = (value?: string | number | null) => (value === undefined ? '' : `${value}`);
-
-const isNonNumericKey = (event: ReactKeyboardEvent<HTMLInputElement>) => {
-    const value = event.key as string;
-    return value.length == 1 ? !/[0-9]/.test(value) : false;
-};
-
-const handleKeydown = (event: ReactKeyboardEvent<HTMLInputElement>) => {
-    if (!event.altKey && !event.ctrlKey && !event.metaKey && isNonNumericKey(event)) {
-        event.preventDefault();
-    }
-};
 
 type NumericInputProps = {
     id: string;
@@ -84,7 +72,6 @@ const NumericInput = ({
                 onChange={handleChange}
                 required={required}
                 placeholder={placeholder}
-                onKeyDown={handleKeydown}
             />
         </EntryWrapper>
     );
